+++ test/res.format.js
@@ -27,7 +27,7 @@ app1.use(function(req, res, next){
 
 app1.use(function(err, req, res, next){
   if (!err.types) throw err;
-  res.send(err.status, 'Supports: ' + err.types.join(', '));
+  res.status(err.status).send('Supports: ' + err.types.join(', '));
 })
 
 var app2 = express();
@@ -41,7 +41,7 @@ app2.use(function(req, res, next){
 });
 
 app2.use(function(err, req, res, next){
-  res.send(err.status, 'Supports: ' + err.types.join(', '));
+  res.status(err.status).send('Supports: ' + err.types.join(', '));
 })
 
 var app3 = express();
@@ -64,7 +64,7 @@ app4.get('/', function(req, res, next){
 });
 
 app4.use(function(err, req, res, next){
-  res.send(err.status, 'Supports: ' + err.types.join(', '));
+  res.status(err.status).send('Supports: ' + err.types.join(', '));
 })
 
 var app5 = express();
@@ -97,7 +97,7 @@ describe('res', function(){
       });
 
       app.use(function(err, req, res, next){
-        res.send(err.status, 'Supports: ' + err.types.join(', '));
+        res.status(err.status).send('Supports: ' + err.types.join(', '));
       });
 
       test(app);
@@ -136,7 +136,7 @@ describe('res', function(){
       });
 
       router.use(function(err, req, res, next){
-        res.send(err.status, 'Supports: ' + err.types.join(', '));
+        res.status(err.status).send('Supports: ' + err.types.join(', '));
       })
 
       app.use(router)